Add debug() and fix a division bug
The division bug was when ProfiniteInteger._div_ was called with
two exact elements of (non-trivial) number fields.
Then we tried to do ``self.modulus / other.value``, which gave an
error since ``self.modulus`` was the zero-ideal. A zero ideal is
not a fractional ideal so no division was implemented.
Fixed it by adding a condition ``if self.modulus.is_zero()``.